requirements.txt
# Core dependencies
pyyaml>=6.0.1
markdown>=3.5.1
python-frontmatter>=1.0.1

# Audio/Video processing
edge-tts>=6.1.9
moviepy>=1.0.3
pillow>=10.2.0
imageio-ffmpeg>=0.4.9

# YouTube API
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0

# CLI and utilities
click>=8.1.7
rich>=13.7.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
flask>=3.0.0
uvicorn>=0.30.0

# Development and testing
pytest>=7.4.0
watchdog>=3.0.0
pytest-playwright>=0.4.0

# Optional advanced features
openai-whisper>=20231117
numpy>=1.24.0,<2.3

# MQTT
paho-mqtt>=1.6.1
